selftests/x86: Consolidate redundant signal helper functions
commit dbd6b649e7d5b66c7fa95a65d67b59cf5b45f0ac upstream.

The x86 selftests frequently register and clean up signal handlers, but
the sethandler() and clearhandler() functions have been redundantly
copied across multiple .c files.

Move these functions to helpers.h to enable reuse across tests,
eliminating around 250 lines of duplicate code.

Converge the error handling by using ksft_exit_fail_msg(), which is
functionally equivalent with err() within the selftest framework.

This change is a prerequisite for the upcoming xstate selftest, which
requires signal handling for registering and cleaning up handlers.

Intel-SIG: commit dbd6b649e7d5 selftests/x86: Consolidate redundant signal helper functions.
DMR ISA APX enabling - HOST

Link: https://lore.kernel.org/r/20250226010731.2456-2-chang.seok.bae@intel.com
[ Quanxian Wang: amend commit log ]
